agent: route compaction/recall through handlers; LiveView replaces ConversationState
Pivots the kernel from owning conversation-state behavior to dispatching
it through handlers. Kernel now holds only the conversation-view role
(renamed `LiveView`, formerly `ConversationState` plus a now-deleted
parallel `InMemoryLiveView` class); compaction and recall are handler
calls that route to whichever extension registers them.

Changes:

- agent-loop: this.conversation retyped to LiveView. Direct method calls
  for compact/search/browse/expand replaced with `this.handlers.call(...)`
  dispatches. Compact handler keeps internal logic only for the manual
  edit cases (rewind/replace); lossy strategies fall through to
  advisors, returning null when none are registered.
- Drops kernel defines that the live-view shape no longer supports:
  conversation:allocate-seq, :reset-for-session, :get-nuclear-entries,
  :get-nuclear-summary, :build-nuclear-block. Also drops the
  loadPriorHistory bootstrap and the eagerNucleateAgent/Tools call sites
  — capture-style strategies subscribe to conversation:message-appended
  for the same work.
- agent:get-compaction-state and context:get-stats drop the nuclearEntries
  / recallArchiveSize fields; neither has external consumers on main.
- tool-protocol, subagent: switch ConversationState parameter types to
  LiveView.
- LiveView.addToolResult signature restored to `string | ImageContent[]`
  so vision-content tool results keep working; substrate-branch's draft
  dropped that branch.
- The LiveView interface is dropped — only one impl exists, no other
  consumers benefit from it. Class plays the role directly.
- multimodal.test.ts: retargeted to LiveView. eagerNucleateTools test
  removed; equivalent coverage will live in the summary-strategy port.
- Delete src/agent/conversation-state.ts (-871 lines).

Transient state: until the summary-strategy extension lands (next commit),
conversation:compact is a no-op for lossy strategies and conversation_recall
returns "no recall strategy registered" — the kernel-as-bus pivot exposes
the surface, the extension fills it.

Refs #174
